backend/requirements.txt
fastapi
uvicorn
openai
anthropic
python-multipart
python-dotenv
pandas
